+++ home-manager/modules/programs/default.nix
@@ -4,6 +4,7 @@
     ./eza.nix
     ./git.nix
     ./gpg.nix
+    ./jujutsu.nix
     ./pms.nix
     ./screen.nix
     ./tridactyl.nix